# docs/swagger.yaml
openapi: 3.0.0
info:
  title: traQ API
  description: traQ v2 API
  version: 2.6.3
  license:
    name: MIT
    url: 'https://github.com/traPtitech/traQ/blob/master/LICENSE'
  contact:
    name: traP
    url: 'https://github.com/traPtitech/traQ'
servers:
  - url: https://q.trap.jp/api/1.0
    description: production environment
  - url: https://traq-dev.tokyotech.org/api/1.0
    description: staging environment
paths:
...
  /users/{userID}:
    parameters:
      - $ref: "#/components/parameters/userIdInPath"
    get:
      tags:
        - user
      operationId: getUser
      description: ユーザーの詳細を取得します。
      security:
        - traqOAuth2:
            - read
            - manage_bot
      responses:
        "200":
          description: |+
            正常に取得できました。
            ユーザー情報を返します。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "404":
          description: +|
            正常に取得できませんでした。
            指定したユーザーは存在しません。

  # PATCH /users/{userID}

  # /users/{userID}/status

  # /users/{userID}/password
...
  /channels/{channelID}:
    parameters:
      - $ref: "#/components/parameters/channelIdInPath"
    get:
      tags:
        - channel
      operationId: getChannel
      description: チャンネルの情報を返します。
      security:
        - traqOAuth2:
            - read
      responses:
        "200":
          description: |+
            正常に取得ができました。
            チャンネルを返します。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "404":
          description: |+
            取得に失敗しました。
            指定したチャンネルは存在しません。
    patch:
      tags:
        - channel
      operationId: editChannel
      description: チャンネルの情報を変更します。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchChannel"
      responses:
        "204":
          description: 正常に変更ができました。
        "400":
          description: 失敗しました。リクエスト内容が不正です。
        "403":
          description: 失敗しました。権限がありません。
        "404":
          description: |+
            変更に失敗しました。
            指定したチャンネルは存在しません。
    delete:
      tags:
        - channel
      operationId: deleteChannel
      description: チャンネルを削除します。
      responses:
        "204":
          description: 削除に成功しました。
        "404":
          description: +|
            削除に失敗しました。
            指定したチャンネルは既に存在しません。
...
  /webhooks:
    get:
      tags:
        - webhook
      operationId: getWebhooks
      description: 自分が作成したwebhookの一覧を取得します。
      security:
        - traqOAuth2:
            - read
            - manage_bot
      responses:
        "200":
          description: 正常に取得できました。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webhook"
    post:
      tags:
        - webhook
      operationId: createWebhooks
      description: webhookを作成します。
      security:
        - traqOAuth2:
            - manage_bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostWebhook"
      responses:
        "201":
          description: 正常に作成できました。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400":
          description: 作成できませんでした。リクエスト内容が不正です。
        "403":
          description: 作成できませんでした。権限がありません。
  /webhooks/{webhookID}:
    parameters:
      - $ref: "#/components/parameters/webhookIdInPath"
    get:
      tags:
        - webhook
      operationId: getWebhook
      description: webhookの詳細を取得します。
      security:
        - traqOAuth2:
            - read
            - manage_bot
      responses:
        "200":
          description: 正常に取得できました。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "403":
          description: 取得できませんでした。権限がありません。
        "404":
          description: 取得できませんでした。指定したwebhookは存在しません。
    patch:
      tags:
        - webhook
      operationId: editWebhook
      description: webhookを修正します。
      security:
        - traqOAuth2:
            - manage_bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchWebhook"
      responses:
        "204":
          description: 正常に修正できました。
        "400":
          description: 修正できませんでした。リクエスト内容が不正です。
        "403":
          description: 削除できませんでした。権限がありません。
        "404":
          description: 修正できませんでした。指定したwebhookは存在しません。
    delete:
      tags:
        - webhook
      operationId: deleteWebhook
      description: webhookを削除します。
      security:
        - traqOAuth2:
            - manage_bot
      responses:
        "204":
          description: 正常に削除できました。
        "403":
          description: 削除できませんでした。権限がありません。
        "404":
          description: 削除できませんでした。指定したwebhookは存在しません。
    post:
      tags:
        - webhook
      operationId: postWebhook
      description: webhookを送信します。
      parameters:
        - in: header
          name: X-TRAQ-Channel-Id
          schema:
            type: string
            format: uuid
          description: 投稿先のチャンネルID(変更する場合)
        - in: header
          name: X-TRAQ-Signature
          schema:
            type: string
          description: リクエストボディシグネチャ。Secretが設定されている場合は必須
        - name: embed
          description: メンション・チャンネルリンクを自動埋め込みする場合に1を指定する
          in: query
          required: false
          schema:
            type: integer
      security: []
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              description: メッセージ文字列
      responses:
        "204":
          description: 正常に送信できました。
        "400":
          description: 正常に送信できませんでした。リクエスト内容が不正です。
        "404":
          description: 正常に送信できませんでした。指定されたwebhookは存在しません。
  /webhooks/{webhookID}/icon:
    parameters:
      - $ref: "#/components/parameters/webhookIdInPath"
    get:
      tags:
        - webhook
      operationId: getWebhookIcon
      description: 指定したWebhookのアイコンを取得します。
      security:
        - traqOAuth2:
            - read
            - manage_bot
      responses:
        "200":
          description: 正常に取得できました。
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
        "404":
          description: 取得できませんでした。指定したWebhookは存在しません。
    put:
      tags:
        - webhook
      operationId: changeWebhookIcon
      description: 指定したwebhookのアイコンを変更します。
      security:
        - traqOAuth2:
            - manage_bot
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              properties:
                file:
                  type: string
                  format: binary
                  description: webhookユーザーのアイコン(1MBまでのpng, jpeg, gif)
            encoding:
              file:
                contentType: image/png, image/jpeg, image/gif
      responses:
        "204":
          description: 正常に変更できました。
        "400":
          description: 変更できませんでした。リクエスト内容が不正です。
        "403":
          description: 変更できませんでした。権限がありません。
        "404":
          description: 修正できませんでした。指定したwebhookは存在しません。

  # /webhooks/{webhookID}/messages
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: r_session
    traqOAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /api/1.0/oauth2/authorize
          tokenUrl: /api/1.0/oauth2/token
          scopes:
            read: 読み取りスコープ
            write: 書き込みスコープ
            manage_bot: bot関連読み書きスコープ
  parameters:
    channelIdInPath:
      name: channelID
      description: 操作の対象となるチャンネルのID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    userIdInPath:
      name: userID
      description: 操作の対象となるユーザーID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    webhookIdInPath:
      name: webhookID
      description: 操作の対象となるWebhookのID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    UUIDs:
      type: array
      description: UUIDの配列
      items:
        type: string
        format: uuid
...
    Channel:
      type: object
      properties:
        channelId:
          type: string
          format: uuid
          description: チャンネルUUID
        name:
          type: string
          description: チャンネル名
          example: yamada
        member:
          $ref: "#/components/schemas/UUIDs"
        parent:
          type: string
          format: uuid
          description: 親の階層のチャンネルUUID
        topic:
          type: string
          description: チャンネルトピック
          example: yamadaチャンネルです！
        children:
          $ref: "#/components/schemas/UUIDs"
        visibility:
          type: boolean
          description: チャンネルの可視状態
        force:
          type: boolean
          description: 強制通知チャンネルか
        private:
          type: boolean
          description: プライベートチャンネルか
        dm:
          type: boolean
          description: ダイレクトメッセージチャンネルか
...
    PatchChannel:
      type: object
      properties:
        name:
          type: string
          description: チャンネル名
          example: gps
        visibility:
          type: boolean
          description: 変更後のvisibility
          example: true
        force:
          type: boolean
          description: 強制通知かどうか
...
    UserStatus:
      type: integer
      enum:
        - 0
        - 1
        - 2
      description: アカウントの状態 (0:停止,1:有効,2:一時停止)
...
    UserDetail:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: ユーザーUUID
        name:
          type: string
          description: traQ ID
        displayName:
          type: string
          description: 表示名
        iconFileId:
          type: string
          format: uuid
          description: アイコンファイルUUID
        bot:
          type: boolean
          description: BOTかどうか
        twitterId:
          type: string
          description: ツイッターID
        lastOnline:
          type: string
          format: date-time
          description: 最終オンライン日時(オンラインに１度もなってない場合はnull)
        isOnline:
          type: boolean
          description: 現在オンラインかどうか
        suspended:
          type: boolean
          description: アカウントが停止中かどうか
        accountStatus:
          $ref: "#/components/schemas/UserStatus"
        tagList:
          $ref: "#/components/schemas/TagList"
...
    Tag:
      type: object
      properties:
        tagId:
          type: string
          format: uuid
          description: タグUUID
        tag:
          type: string
          description: タグ文字列
        isLocked:
          type: boolean
          description: タグがロックされているかどうか
        createdAt:
          type: string
          format: date-time
          description: タグ付与日時
        updatedAt:
          type: string
          format: date-time
          description: タグ更新日時
    TagList:
      type: array
      items:
        $ref: "#/components/schemas/Tag"
...
    Webhook:
      type: object
      properties:
        webhookId:
          type: string
          format: uuid
          description: Webhook UUID
        botUserId:
          type: string
          format: uuid
          description: WebhookユーザーのUUID
        displayName:
          type: string
          description: Webhookユーザー表示名
        description:
          type: string
          description: Webhook説明
        secure:
          type: boolean
          description: Secure Webhookかどうか
        channelId:
          type: string
          format: uuid
          description: デフォルト投稿先チャンネルUUID
        creatorId:
          type: string
          format: uuid
          description: Webhook作成者UUID
        createdAt:
          type: string
          format: date-time
          description: Webhook作成日時
        updatedAt:
          type: string
          format: date-time
          description: Webhook更新日時
    PostWebhook:
      type: object
      required:
        - name
        - description
        - channelId
      properties:
        name:
          type: string
          maxLength: 32
          description: webhookユーザーの表示名
        description:
          type: string
          description: webhookの説明
        channelId:
          type: string
          format: uuid
          description: デフォルトの投稿先チャンネル(パブリックチャンネルのみ)
        secret:
          type: string
          description: webhookシークレット
    PatchWebhook:
      type: object
      properties:
        name:
          type: string
          maxLength: 32
          description: webhookユーザーの表示名
        description:
          type: string
          description: webhookの説明
        channelId:
          type: string
          format: uuid
          description: デフォルトの投稿先チャンネル(パブリックチャンネルのみ)
        secret:
          type: string
          description: webhookシークレット
        creatorId:
          type: string
          format: uuid
          description: 移譲先のユーザーUUID
...
security:
  - cookieAuth: []